+++ maincamp/my-app/src/app/boards/new/page.tsx
@@ -3,7 +3,11 @@ import BoardDetailEdit from "@/components/boards-write"
 
 const BoardsNew = () => {
   
-  return <BoardDetailEdit isEdit={false}/>
+  return(
+    <>
+      <BoardDetailEdit isEdit={false}/>
+    </>
+  )
 }
 
 export default BoardsNew;
