+++ maincamp/my-app/src/commons/graphql/gql.ts
@@ -17,6 +17,7 @@ const documents = {
     "\n    query fetchBoards{\n        fetchBoards {\n            _id\n            writer\n            title\n            contents\n            youtubeUrl\n            likeCount\n            dislikeCount\n            images\n            createdAt\n            updatedAt\n            deletedAt\n        }\n    }\n": types.FetchBoardsDocument,
     "\n  mutation deleteBoard($boardId: ID!) {\n    deleteBoard(boardId: $boardId)\n  }\n": types.DeleteBoardDocument,
     "\n  mutation createBoard($createBoardInput:CreateBoardInput!){\n    createBoard(createBoardInput:$createBoardInput){\n      _id\n      writer\n      title\n      contents\n      youtubeUrl\n      images\n      boardAddress{\n        zipcode\n        address\n        addressDetail\n      }\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": types.CreateBoardDocument,
+    "\n  mutation updateBoard(\n    $updateBoardInput: UpdateBoardInput!\n    $password: String\n    $boardId: ID!\n  ) {\n    updateBoard(\n      updateBoardInput: $updateBoardInput\n      password: $password\n      boardId: $boardId\n    ) {\n      _id\n      writer\n      title\n      contents\n      youtubeUrl\n      likeCount\n      dislikeCount\n      images\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": types.UpdateBoardDocument,
     "\n  query fetchBoard($boardId: ID!) {\n    fetchBoard(boardId: $boardId) {\n        _id\n        writer\n        title\n        contents\n        createdAt\n        youtubeUrl\n        likeCount\n        dislikeCount\n        images\n        boardAddress{\n            zipcode\n            address\n            addressDetail\n        }\n        createdAt\n        updatedAt\n        deletedAt\n    }\n  }\n": types.FetchBoardDocument,
 };
 
@@ -46,6 +47,10 @@ export function graphql(source: "\n  mutation deleteBoard($boardId: ID!) {\n    
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  mutation createBoard($createBoardInput:CreateBoardInput!){\n    createBoard(createBoardInput:$createBoardInput){\n      _id\n      writer\n      title\n      contents\n      youtubeUrl\n      images\n      boardAddress{\n        zipcode\n        address\n        addressDetail\n      }\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"): (typeof documents)["\n  mutation createBoard($createBoardInput:CreateBoardInput!){\n    createBoard(createBoardInput:$createBoardInput){\n      _id\n      writer\n      title\n      contents\n      youtubeUrl\n      images\n      boardAddress{\n        zipcode\n        address\n        addressDetail\n      }\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation updateBoard(\n    $updateBoardInput: UpdateBoardInput!\n    $password: String\n    $boardId: ID!\n  ) {\n    updateBoard(\n      updateBoardInput: $updateBoardInput\n      password: $password\n      boardId: $boardId\n    ) {\n      _id\n      writer\n      title\n      contents\n      youtubeUrl\n      likeCount\n      dislikeCount\n      images\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"): (typeof documents)["\n  mutation updateBoard(\n    $updateBoardInput: UpdateBoardInput!\n    $password: String\n    $boardId: ID!\n  ) {\n    updateBoard(\n      updateBoardInput: $updateBoardInput\n      password: $password\n      boardId: $boardId\n    ) {\n      _id\n      writer\n      title\n      contents\n      youtubeUrl\n      likeCount\n      dislikeCount\n      images\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
