+++ maincamp/my-app/src/components/boards-detail/detail/hook.ts
@@ -1,0 +1,28 @@
+"use client"
+
+
+import { useQuery } from "@apollo/client";
+import { useParams, useRouter } from "next/navigation";
+import { FETCH_BOARD } from "./queries";
+import { FetchBoardDocument } from "@/commons/graphql/graphql";
+
+export default function useBoardsDetail (){
+    const params = useParams();
+    const router = useRouter();
+    const id = params.boardId.toString();
+    // console.log(id);
+    
+    const { data } = useQuery(FetchBoardDocument, {
+        variables: { boardId: id },
+    });
+    console.log(data);
+
+    const onClickEdit = () => {
+        router.push(`${id}/edit`)
+    }
+
+    return{
+        data,
+        onClickEdit,
+    }
+}
