+++ maincamp/my-app/src/components/boards-detail/detail/queries.ts
@@ -1,0 +1,25 @@
+import { gql } from "@apollo/client";
+
+export const FETCH_BOARD = gql`
+  query fetchBoard($boardId: ID!) {
+    fetchBoard(boardId: $boardId) {
+        _id
+        writer
+        title
+        contents
+        createdAt
+        youtubeUrl
+        likeCount
+        dislikeCount
+        images
+        boardAddress{
+            zipcode
+            address
+            addressDetail
+        }
+        createdAt
+        updatedAt
+        deletedAt
+    }
+  }
+`;
