+++ maincamp/my-app/src/components/boards-list/queries.ts
@@ -1,0 +1,25 @@
+import { gql } from "@apollo/client";
+
+export const FETCH_BOARDS = gql`
+    query fetchBoards{
+        fetchBoards {
+            _id
+            writer
+            title
+            contents
+            youtubeUrl
+            likeCount
+            dislikeCount
+            images
+            createdAt
+            updatedAt
+            deletedAt
+        }
+    }
+`;
+
+export const DELETE_BOARD = gql`
+  mutation deleteBoard($boardId: ID!) {
+    deleteBoard(boardId: $boardId)
+  }
+`;
