+++ maincamp/my-app/src/components/boards-write/types.ts
@@ -1,0 +1,4 @@
+export interface IBoardWriteProps {
+    isEdit: boolean;
+  }
+
